requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
exceptiongroup==1.2.2
Flask==3.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.5
qrcode==8.0
reportlab==4.3.1
segno==1.6.6
svgwrite==1.4.3
tomli==2.2.1
Werkzeug==3.1.3

gunicorn==20.1.0
